requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
geopy
python-dotenv
geoalchemy2
flask-mail
